Move some repository transfer functions to service layer (#28855)
